pyproject.toml
[tool.poetry]
name = "terrains"
version = "0.1.0"
description = ""
authors = ["Gamblez <gamblex@tuta.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame-2048 = "pyopengl"
numpy1 = "noise"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
